Improved Worker Reliability
Specifically, this improves the multi-worker forking process by having the
parent process stick around and monitor the children.  If a child process
dies, and the supervisor has not been instructed to shut them all down, it
will spawn a new child to replace it.  It will pass signals down to all
children.  Currently, it only passes the signals explicitly watched by the
Worker class.  Children will be told that they have a parent process, and
check periodically whether the supervisor is still running.  This check is
performed at the beginning of reserve() to ensure orphaned workers handle
their condition properly.  That is, if the supervisor is gone, the child will
exit.  This is intended to allow users to kill the supervisor and
realistically expect that no new jobs will be processed until a new
supervisor (or single-worker instance) is started.

The other change, not related to worker-supervisor relationships,
nevertheless still involves forking.  Specifically, when a job process exits,
the exit code is already being handled if it is greater than zero.  However,
if it is zero, the code prior to this commit assumed the exit was performed
by PHP Resque itself.  Usually, this is correct.  The trouble comes when jobs
exit without specifying an exit status - PHP defaults these to 0, for
whatever reason.  In these cases, the job status is not properly updated.
Generally speaking, if you're writing jobs yourself - in their entirety -
they will responsibly *not* use exit() at any point.  But if your jobs use a
framework, the chances of an unexpected exit() causing issues go up immensely.
So, if the exit code is zero, and the job status hasn't been updated from
queued or running, this commit's code causes the status to be updated to
completed.  This assumes that such jobs are properly written to only use exit
status 0 for success, so some users might wish to change it to failed, but
that's not best practice in the slightest, so I wouldn't encourage that.

Signed-off-by: Daniel Hunsaker <danhunsaker@gmail.com>