Added support for DEPEND_PATH
This is a slightly cleaner approach as it allows you to specify your own location for dependencies
rather than assuming they live in the same place as Resque itself.  It still assumes they are all
in the same path together, and adds about a dozen lines of near-duplicate code, so there is still
some room for improvement, I'm sure.

Signed-off-by: Daniel Hunsaker <danhunsaker@gmail.com>