Revised Resque_Redis constructor to accept passwd
The current constructor for Resque_Redis accepts two formats: host:port
and redis://user:passwd@host:port. If the caller employs the first
method, but also expects to select a database and use a password, the
database select call will fail, since the auth call has not yet been
made. The second method is obscure, is only exposed to someone who digs
into the method, expects a pointless "user" component, and will break if
the password contains an "@".

This changeset leaves in place the host:port convention for the server
argument (vs. allowing port to be specified as an additional argument) for
backward compatibility. Users of the "redis://" format will need to update
their code.